매칭 응답 chatRoomId 추가, dto 통합
