--- src/matching/dto/Patch-matching.request.ts
@@ -1,13 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsString, IsIn } from 'class-validator';
-
-export class PatchMatchingRequest {
-  @ApiProperty({
-    example: 'accept',
-    enum: ['accept', 'reject'],
-    description: '수락 또는 거절',
-  })
-  @IsString()
-  @IsIn(['accept', 'reject'])
-  requestStatus: 'accept' | 'reject';
-}
